+++ src/main/kotlin/com/bugsnag/android/gradle/BugsnagPlugin.kt
@@ -340,9 +340,25 @@ class BugsnagPlugin : Plugin<Project> {
             project.rootProject.allprojects { subProj ->
                 val defaultNodeModulesDir = File("${subProj.rootDir}/../node_modules")
                 val nodeModulesDir = bugsnag.nodeModulesDir.getOrElse(defaultNodeModulesDir)
+                if (!nodeModulesDir.exists()) {
+                    throw StopExecutionException(
+                        "Cannot find node_modules directory at: ${nodeModulesDir.absolutePath} " +
+                            "To set this to the correct path manually, please see: " +
+                            "https://docs.bugsnag.com/build-integrations/gradle/#custom-node_modules-directory"
+                    )
+                }
+
+                val bugsnagModuleDir = File(nodeModulesDir, "@bugsnag/react-native/android")
+                if (!bugsnagModuleDir.exists()) {
+                    throw StopExecutionException(
+                        "Cannot find the @bugsnag/react-native module in your node_modules directory. " +
+                            "Manual installation instructions can be found here: " +
+                            "https://docs.bugsnag.com/platforms/react-native/react-native/manual-setup/#installation"
+                    )
+                }
 
                 subProj.repositories.maven { repo ->
-                    repo.setUrl("$nodeModulesDir/@bugsnag/react-native/android")
+                    repo.setUrl(bugsnagModuleDir.toString())
                 }
             }
         }
@@ -475,12 +491,22 @@ class BugsnagPlugin : Plugin<Project> {
         val rnTaskName = "bundle${variant.name.capitalize()}JsAndAssets"
         val rnTask: Task = project.tasks.findByName(rnTaskName) ?: return null
         val rnSourceMap = findReactNativeSourcemapFile(project, variant)
-        val rnBundle = BugsnagUploadJsSourceMapTask.findReactNativeTaskArg(rnTask, "--bundle-output")
+        var rnBundle = BugsnagUploadJsSourceMapTask.findReactNativeTaskArg(rnTask, "--bundle-output")
         val dev = BugsnagUploadJsSourceMapTask.findReactNativeTaskArg(rnTask, "--dev")
 
         if (rnBundle == null || dev == null) {
             project.logger.error("Bugsnag: unable to upload JS sourcemaps. Please enable sourcemap + bundle output.")
             return null
+        }
+
+        val enabledHermes = BugsnagUploadJsSourceMapTask.isHermesEnabled(project)
+        if (enabledHermes) {
+            rnBundle = BugsnagUploadJsSourceMapTask.rescueReactNativeSourceBundle(
+                rnTask,
+                rnBundle,
+                project,
+                output
+            )
         }
 
         return BugsnagUploadJsSourceMapTask.register(project, taskName) {